misc/HtmlTestSuite/results/test-0018.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>Allow trailing empty entry in value list</title>
<meta charset="utf-8"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<link href="https://svgwg.org/specs/animations/#ValuesAttribute" rel="help"/>
<link href="mailto:edvardt@opera.com" rel="author" title="Edvard Thörnros"/>

</head><body><svg xmlns="http://www.w3.org/2000/svg" viewbox="0 0 100 100">
  <path fill="#1EB287">
    <animate attributename="d" dur="1s" repeatcount="indefinite" values="M 0,0 L 50, 0 L 100,100 Z;&#xA;                     M 100,0 L 0, 100 L 50, 0 Z; "/>
  </path>
</svg>

<script>
  async_test(function(t) {
    window.onload = t.step_func(function() {
      window.requestAnimationFrame(t.step_func_done(function() {
        let path = document.querySelector("path");
        assert_not_equals(window.getComputedStyle(path).d, "none");
      }));
    });
  });
</script>
</body></html>
